TEAM-6: check permision user team

See merge request t-bcards/asap/c-sendmail/api!393

+++ app/Http/Controllers/Api/FooterTemplateController.php
@@ -169,6 +169,9 @@ class FooterTemplateController extends AbstractRestAPIController
 
     public function editMyFooterTemplate(UpdateMyFooterTemplateRequest $request, $id)
     {
+        if (!Gate::allows('permission', config('api.footer.edit'))) {
+            return $this->sendJsonResponse(false, 'You need to buy platform/add-on', ['data' => $this->getPlatformByPermission(config('api.footer.edit'))], 403);
+        }
         $model = $this->myService->showMyFooterTemplate($id);
 
         if ($request->get('active_by_uuid')) {
@@ -183,6 +186,9 @@ class FooterTemplateController extends AbstractRestAPIController
 
     public function destroyMyFooterTemplate($id)
     {
+        if (!Gate::allows('permission', config('api.footer.delete'))) {
+            return $this->sendJsonResponse(false, 'You need to buy platform/add-on', ['data' => $this->getPlatformByPermission(config('api.footer.delete'))], 403);
+        }
         $model = $this->myService->showMyFooterTemplate($id);
         if (!$model->active_by_uuid) {
             $this->service->destroy($id);
